-【add】多行业预约下单，迁移我要加入管理后台

+++ api-mall/src/main/java/com/hailu/cloud/api/mall/module/multiindustry/service/StoreInformationService.java
@@ -19,5 +19,5 @@ public interface StoreInformationService {
      * @param id
      * @return
      */
-    StoreInformation findStoreInformation(Integer id);
+    StoreInformation findStoreInformation(Long id);
 }
