+++ api-mall/src/main/java/com/hailu/cloud/api/mall/module/multiindustry/service/impl/StoreInformationImpl.java
@@ -25,7 +25,7 @@ public class StoreInformationImpl implements StoreInformationService {
     }
 
     @Override
-    public StoreInformation findStoreInformation(Integer id) {
+    public StoreInformation findStoreInformation(Long id) {
         return storeInformationMapper.findStoreInformation(id);
     }
 }
